moving necessary casting code to header and doing adjustments to accept logical types in the result
